% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{semislv}
\alias{semislv}
\title{Solve Semi-parametric estimation by implicit profiling}
\usage{
semislv(
  theta,
  lambda,
  Phi_fn,
  Psi_fn,
  jac = list(),
  intermediates = list(),
  method = c("iterative", "implicit"),
  diy = FALSE,
  control = list(max_iter = 100, tol = 0.001),
  ...
)
}
\arguments{
\item{theta}{the initial value of parametric part}

\item{lambda}{the initial value of non-parametric part}

\item{Phi_fn}{the equation function highly relevant to the parametric part}

\item{Psi_fn}{the equation function highly relevant to the non-parametric part}

\item{jac}{a list containing some of deterivate info of Phi_der_theta_fn, Psi_der_theta_fn, Phi_der_lambda_fn, Psi_der_lambda_fn,}

\item{method}{"implicit" or "iterative"}

\item{diy}{a bool value to decide to parse user designed function}

\item{control}{a list like list(max_iter = 100, tol = 1e-3) to control the early stop}

\item{...}{static parameter for Phi_fn, Psi_fn. Diy execution function.}
}
\value{
A save space containing final iteration result and iteration path
}
\description{
Solve Semi-parametric estimation by implicit profiling
}
\examples{
Phi_fn <- function(theta, lambda, alpha) 2 * theta + alpha * lambda
Psi_fn <- function(theta, lambda, alpha) 2 * lambda + alpha * theta
res <- semislv(1, 1, Phi_fn, Psi_fn, method = "implicit", alpha = 1)
}
